Improve load time of anomalies page by 2x
